+++ uvdat/core/frame_previews/raster_style.py
