+++ uvdat/core/models/styles.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 
-from uvdat.core.frame_preview_types import FramePreviewData
+from uvdat.core.frame_previews.types import FramePreviewData
 
 from .colormap import Colormap
 from .layer import Layer
